+++ packages/react/src/index.ts
@@ -7,14 +7,14 @@ export {
   AccordionItem,
   type AccordionItemIndicatorProps,
   type AccordionItemProps,
-  type AccordionProps
+  type AccordionProps,
 } from "@jamsr-ui/accordion";
 export { Alert, type AlertProps } from "@jamsr-ui/alert";
 export {
   Autocomplete,
   AutocompleteItem,
   type AutocompleteItemProps,
-  type AutocompleteProps
+  type AutocompleteProps,
 } from "@jamsr-ui/autocomplete";
 export { Avatar, type AvatarProps } from "@jamsr-ui/avatar";
 export { Button, type ButtonProps } from "@jamsr-ui/button";
@@ -26,7 +26,7 @@ export {
   type CardContentProps,
   type CardFooterProps,
   type CardHeaderProps,
-  type CardProps
+  type CardProps,
 } from "@jamsr-ui/card";
 export {
   AreaChart,
@@ -45,7 +45,7 @@ export { Checkbox, type CheckboxProps } from "@jamsr-ui/checkbox";
 export { Chip, type ChipProps } from "@jamsr-ui/chip";
 export {
   CircularProgress,
-  type CircularProgressProps
+  type CircularProgressProps,
 } from "@jamsr-ui/circular-progress";
 export {
   Collapsible,
@@ -55,18 +55,18 @@ export {
   type CollapsibleContentProps,
   type CollapsibleProps,
   type CollapsibleStateProps,
-  type CollapsibleTriggerProps
+  type CollapsibleTriggerProps,
 } from "@jamsr-ui/collapsible";
 export {
   Confirmation,
   useConfirmation,
-  type ConfirmationProps
+  type ConfirmationProps,
 } from "@jamsr-ui/confirmation";
 export { UIProvider } from "@jamsr-ui/core";
 export {
   DataTable,
   type ColumnDef,
-  type DataTableProps
+  type DataTableProps,
 } from "@jamsr-ui/data-table";
 export {
   Dialog,
@@ -78,13 +78,13 @@ export {
   type DialogContentProps,
   type DialogFooterProps,
   type DialogHeaderProps,
-  type DialogProps
+  type DialogProps,
 } from "@jamsr-ui/dialog";
 export { Divider, type DividerProps } from "@jamsr-ui/divider";
 export {
   Sortable,
   type SortableItemProps,
-  type UniqueIdentifier
+  type UniqueIdentifier,
 } from "@jamsr-ui/dnd";
 export {
   Drawer,
@@ -92,18 +92,18 @@ export {
   DrawerFooter,
   DrawerHeader,
   type DrawerBodyProps,
-  type DrawerProps
+  type DrawerProps,
 } from "@jamsr-ui/drawer";
 export { Editor, type EditorProps } from "@jamsr-ui/editor";
 export {
   FileUploadMulti,
   type FileUploadMultiProps,
-  type FileUploadMultiState
+  type FileUploadMultiState,
 } from "@jamsr-ui/file-upload-multi";
 export {
   FileUploadSingle,
   type FileUploadError,
-  type FileUploadSingleProps
+  type FileUploadSingleProps,
 } from "@jamsr-ui/file-upload-single";
 export { Header, type HeaderProps } from "@jamsr-ui/header";
 export { IconButton, type IconButtonProps } from "@jamsr-ui/icon-button";
@@ -111,14 +111,14 @@ export { Input, type InputProps } from "@jamsr-ui/input";
 export { Kbd, type KbdProps } from "@jamsr-ui/kbd";
 export {
   LinearProgress,
-  type LinearProgressProps
+  type LinearProgressProps,
 } from "@jamsr-ui/linear-progress";
 export { Link, type LinkProps } from "@jamsr-ui/link";
 export {
   Menu,
   MenuItem,
   type MenuItemProps,
-  type MenuProps
+  type MenuProps,
 } from "@jamsr-ui/menu";
 export { OtpInput, type OtpInputProps } from "@jamsr-ui/otp-input";
 export { Popover, type PopoverProps } from "@jamsr-ui/popover";
@@ -126,7 +126,7 @@ export {
   Radio,
   RadioGroup,
   type RadioGroupProps,
-  type RadioProps
+  type RadioProps,
 } from "@jamsr-ui/radio";
 export { Rating, type RatingProps } from "@jamsr-ui/rating";
 export { Repeater } from "@jamsr-ui/repeater";
@@ -144,14 +144,14 @@ export {
   RHFSelect,
   RHFSwitch,
   RHFTagsInput,
-  RHFTextarea
+  RHFTextarea,
 } from "@jamsr-ui/rhf";
 export { Ripple } from "@jamsr-ui/ripple";
 export {
   Select,
   SelectItem,
   type SelectItemProps,
-  type SelectProps
+  type SelectProps,
 } from "@jamsr-ui/select";
 export {
   Sidebar,
@@ -171,14 +171,14 @@ export {
   type SidebarMenuItemButtonProps,
   type SidebarMenuItemProps,
   type SidebarMenuProps,
-  type SidebarProps
+  type SidebarProps,
 } from "@jamsr-ui/sidebar";
 export { Skeleton, type SkeletonProps } from "@jamsr-ui/skeleton";
 export { Stepper, type StepperProps } from "@jamsr-ui/stepper";
 export {
   UIStylesProvider,
   useUIStyle,
-  type UIStylesType
+  type UIStylesType,
 } from "@jamsr-ui/styles";
 export { Switch, type SwitchProps } from "@jamsr-ui/switch";
 export { Tab, Tabs, type TabProps, type TabsProps } from "@jamsr-ui/tab";
@@ -194,11 +194,10 @@ export {
   type TableColumnProps,
   type TableHeaderProps,
   type TableProps,
-  type TableRowProps
+  type TableRowProps,
 } from "@jamsr-ui/table";
 export { TagsInput, type TagsInputProps } from "@jamsr-ui/tags-input";
 export { Text, textVariants, type TextProps } from "@jamsr-ui/text";
 export { Textarea, type TextareaProps } from "@jamsr-ui/textarea";
 export { toast, ToastProvider, type ToastProps } from "@jamsr-ui/toast";
 export { Tooltip, type TooltipProps } from "@jamsr-ui/tooltip";
-