+++ apps/www/CHANGELOG.md
@@ -1,5 +1,18 @@
 # web
 
+## 0.1.204
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/react@0.0.222
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/shared-icons@0.0.68
+  - @jamsr-ui/theme@0.0.79
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.1.203
 
 ### Patch Changes