+++ packages/components/avatar/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jamsr-ui/avatar
 
+## 0.0.82
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/shared-icons@0.0.68
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.81
 
 ### Patch Changes