+++ packages/components/circular-progress/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/circular-progress
 
+## 0.0.10
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.9
 
 ### Patch Changes