+++ packages/components/data-table/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @jamsr-ui/data-table
 
+## 0.0.76
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/button@0.0.87
+  - @jamsr-ui/circular-progress@0.0.10
+  - @jamsr-ui/icon-button@0.0.12
+  - @jamsr-ui/linear-progress@0.0.9
+  - @jamsr-ui/repeater@0.0.45
+  - @jamsr-ui/select@0.0.106
+  - @jamsr-ui/skeleton@0.0.77
+  - @jamsr-ui/table@0.0.61
+  - @jamsr-ui/text@0.0.4
+  - @jamsr-ui/shared-icons@0.0.68
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.75
 
 ### Patch Changes