+++ packages/components/drawer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/drawer
 
+## 1.0.83
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/icon-button@0.0.12
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/shared-icons@0.0.68
+  - @jamsr-ui/utils@0.0.73
+
 ## 1.0.82
 
 ### Patch Changes