+++ packages/components/file-upload-single/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/file-upload-single
 
+## 0.0.109
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/circular-progress@0.0.10
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/shared-icons@0.0.68
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.108
 
 ### Patch Changes