+++ packages/components/kbd/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/kbd
 
+## 0.0.14
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.13
 
 ### Patch Changes