+++ packages/components/menu/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @jamsr-ui/menu
 
+## 0.0.104
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/shared-icons@0.0.68
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.103
 
 ### Patch Changes