+++ packages/components/otp-input/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jamsr-ui/otp-input
 
+## 0.0.80
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.79
 
 ### Patch Changes