+++ packages/components/popover/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jamsr-ui/popover
 
+## 0.0.83
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.82
 
 ### Patch Changes