+++ packages/components/radio/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @jamsr-ui/radio
 
+## 0.0.88
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.87
 
 ### Patch Changes