+++ packages/components/switch/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @jamsr-ui/switch
 
+## 0.0.90
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/text@0.0.4
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.89
 
 ### Patch Changes