+++ packages/components/table/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/table
 
+## 0.0.61
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.60
 
 ### Patch Changes