+++ packages/components/tags-input/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @jamsr-ui/tags-input
 
+## 0.0.48
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/input@0.0.110
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/chip@0.0.95
+  - @jamsr-ui/hooks@0.0.55
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.47
 
 ### Patch Changes