+++ packages/components/text/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/typography
 
+## 0.0.4
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.3
 
 ### Patch Changes