+++ packages/components/tooltip/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @jamsr-ui/tooltip
 
+## 0.0.78
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.77
 
 ### Patch Changes