+++ packages/hooks/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @jamsr-ui/hooks
 
+## 0.0.55
+
+### Patch Changes
+
+- fix treeshake
+
 ## 0.0.54
 
 ### Patch Changes