+++ packages/react/CHANGELOG.md
@@ -1,5 +1,62 @@
 # @jamsr-ui/react
 
+## 0.0.222
+
+### Patch Changes
+
+- fix treeshake
+- Updated dependencies
+  - @jamsr-ui/editor@0.0.120
+  - @jamsr-ui/input@0.0.110
+  - @jamsr-ui/styles@0.0.17
+  - @jamsr-ui/accordion@0.0.90
+  - @jamsr-ui/alert@0.0.97
+  - @jamsr-ui/autocomplete@0.0.119
+  - @jamsr-ui/avatar@0.0.82
+  - @jamsr-ui/button@0.0.87
+  - @jamsr-ui/card@0.0.92
+  - @jamsr-ui/charts@0.0.5
+  - @jamsr-ui/checkbox@0.0.83
+  - @jamsr-ui/chip@0.0.95
+  - @jamsr-ui/circular-progress@0.0.10
+  - @jamsr-ui/collapsible@0.0.3
+  - @jamsr-ui/confirmation@0.0.79
+  - @jamsr-ui/data-table@0.0.76
+  - @jamsr-ui/dialog@0.0.97
+  - @jamsr-ui/divider@0.0.79
+  - @jamsr-ui/dnd@0.0.18
+  - @jamsr-ui/drawer@1.0.83
+  - @jamsr-ui/file-upload-multi@0.0.106
+  - @jamsr-ui/file-upload-single@0.0.109
+  - @jamsr-ui/header@0.0.85
+  - @jamsr-ui/icon-button@0.0.12
+  - @jamsr-ui/kbd@0.0.14
+  - @jamsr-ui/linear-progress@0.0.9
+  - @jamsr-ui/link@0.0.79
+  - @jamsr-ui/menu@0.0.104
+  - @jamsr-ui/otp-input@0.0.80
+  - @jamsr-ui/popover@0.0.83
+  - @jamsr-ui/radio@0.0.88
+  - @jamsr-ui/rating@0.0.41
+  - @jamsr-ui/repeater@0.0.45
+  - @jamsr-ui/rhf@0.0.64
+  - @jamsr-ui/ripple@0.0.81
+  - @jamsr-ui/select@0.0.106
+  - @jamsr-ui/sidebar@0.0.88
+  - @jamsr-ui/skeleton@0.0.77
+  - @jamsr-ui/stepper@0.0.19
+  - @jamsr-ui/switch@0.0.90
+  - @jamsr-ui/tab@0.0.102
+  - @jamsr-ui/table@0.0.61
+  - @jamsr-ui/tags-input@0.0.48
+  - @jamsr-ui/text@0.0.4
+  - @jamsr-ui/textarea@0.0.110
+  - @jamsr-ui/toast@0.0.39
+  - @jamsr-ui/tooltip@0.0.78
+  - @jamsr-ui/core@0.0.47
+  - @jamsr-ui/theme@0.0.79
+  - @jamsr-ui/utils@0.0.73
+
 ## 0.0.221
 
 ### Patch Changes